refactor : 로또 발생 시 필요한 기능을 CreateLottos 클래스로 분리
