+++ src/main/kotlin/lotto/Lotto.kt
@@ -1,10 +1,20 @@
 package lotto
 
-class Lotto(private val numbers: List<Int>) {
+//Lotto 클래스에 들어갈 기능
+/*
+
+ */
+
+class Lotto(private val numbers: List<Int>){
     init {
         require(numbers.size == 6)
     }
 
     // TODO: 추가 기능 구현
+    companion object{
+
+    }
+
+
 
 }
